Move common utility `setFocus()` to separate file
