+++ packages/nhsuk-frontend/src/nhsuk/common/index.mjs
@@ -43,62 +43,6 @@ export function getBreakpoint(name) {
 }
 
 /**
- * Move focus to element
- *
- * Sets tabindex to -1 to make the element programmatically focusable,
- * but removes it on blur as the element doesn't need to be focused again.
- *
- * Original code taken from GDS (Government Digital Service)
- * {@link https://github.com/alphagov/govuk-frontend}
- *
- * @template {HTMLElement} FocusElement
- * @param {FocusElement} $element - HTML element
- * @param {object} [options] - Handler options
- * @param {function(this: FocusElement): void} [options.onBeforeFocus] - Callback before focus
- * @param {function(this: FocusElement): void} [options.onBlur] - Callback on blur
- */
-export function setFocus($element, options = {}) {
-  const isFocusable = $element.getAttribute('tabindex')
-
-  if (!isFocusable) {
-    $element.setAttribute('tabindex', '-1')
-  }
-
-  /**
-   * Handle element focus
-   */
-  function onFocus() {
-    $element.removeEventListener('focus', onFocus)
-    $element.addEventListener('blur', onBlur)
-  }
-
-  /**
-   * Handle element blur
-   */
-  function onBlur() {
-    $element.removeEventListener('blur', onBlur)
-
-    if (options.onBlur) {
-      options.onBlur.call($element)
-    }
-
-    if (!isFocusable) {
-      $element.removeAttribute('tabindex')
-    }
-  }
-
-  // Add listener to reset element on blur, after focus
-  $element.addEventListener('focus', onFocus)
-
-  // Focus element
-  if (options.onBeforeFocus) {
-    options.onBeforeFocus.call($element)
-  }
-
-  $element.focus()
-}
-
-/**
  * Checks if component is already initialised
  *
  * @param {Element} $root - HTML element to be checked
@@ -164,6 +108,7 @@ export function formatErrorMessage(Component, message) {
 
 export * from './closest-attribute-value.mjs'
 export * from './nhsuk-frontend-version.mjs'
+export * from './set-focus.mjs'
 
 /**
  * @import { CompatibleClass } from '../component.mjs'
